(fix): don't reapply tags every run when torrent (#824)
is stalled but tag_stalled_torrents is set to false

<!--
Thanks for opening a PR! Your contribution is much appreciated.
In order to make sure your PR is handled as smoothly as possible we
request that you follow the checklist sections below.
Choose the right checklist for the change that you're making:
-->

## Description

If `tag_stalled_torrents` is set to `false` and a torrent is stalled,
already correctly tagged torrents are getting their tags re-applied each
time we run `qbit_manage`.

**Reason**
Condition checks only for stalled torrent and if stalled tag is applied,
but does not check if `tag_stalled_torrents: true`.

**Fix**
This fix adds the check for `tag_stalled_torrents: true`

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added or updated the docstring for new or existing methods
- [x] I have modified this PR to merge to the develop branch