[Bug]: Using -2 as a max ratio target causes qbit_manage to repeatedly update a torrent's share limits (#832)
<!--
Thanks for opening a PR! Your contribution is much appreciated.
In order to make sure your PR is handled as smoothly as possible we
request that you follow the checklist sections below.
Choose the right checklist for the change that you're making:
-->

## Description

I **think** the issue I described in #831 is happening because torrents
contain two fields that appear to be getting mixed up: `max_ratio` and
`ratio_limit`. I believe the `ratio_limit` is what the specific torrent
is set to, and the `max_ratio` is the effective ratio. So for example a
torrent with no max ratio set and a global limit of 10 in effect shows
in the JSON data as:

```
"max_ratio": 10,
"ratio_limit": -2,
```

Therefore I think qbit_manage should be comparing against `ratio_limit`
rather than `max_ratio` before it makes its changes. I have tested that
this makes my config in #831 work as expected, but I am new to this
codebase and would appreciate a developer's review of this for anything
I may be missing here.

Fixes #831

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added or updated the docstring for new or existing methods
- [x] I have modified this PR to merge to the develop branch

---------

Co-authored-by: bobokun <12660469+bobokun@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>