Add custom hostname option (#929)
<!--
Thanks for opening a PR! Your contribution is much appreciated.
In order to make sure your PR is handled as smoothly as possible we
request that you follow the checklist sections below.
Choose the right checklist for the change that you're making:
-->

## Description

Add `--host` / `-H` CLI argument to allow users to specify the host
address the UI should bind to. This allows for more flexibility so
people can bind only to `127.0.0.1` for example instead of the default
of `0.0.0.0` which exposes all interfaces.

Fixes # N/A

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update


## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] (N/A) I have commented my code, particularly in hard-to-understand
areas
- [ ] (N/A) I have added or updated the docstring for new or existing
methods
- [x] I have modified this PR to merge to the develop branch

---------

Co-authored-by: bobokun <12660469+bobokun@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>